requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
